feat: 매뉴얼 섹션 컴포넌트 추가 및 이벤트 리스너 통합
`ManualSection.ts` 파일을 새로 생성하여 매뉴얼 토글 및 오버레이 기능을 통합했습니다. `main.basic.ts`에서 이벤트 리스너를 설정하는 로직을 간소화하고, `App.ts`에서 매뉴얼 섹션을 추가하여 코드의 가독성과 유지보수성을 향상시켰습니다. 또한, 기존의 매뉴얼 관련 코드 일부를 제거하여 구조를 개선했습니다.
