+++ index.html
@@ -42,6 +42,6 @@
     >
       <!-- Content will be dynamically generated here -->
     </div>
-    <script type="module" src="/src/main.original.js"></script>
+    <script type="module" src="/src/basic/main.basic.ts"></script>
   </body>
 </html>
